+++ backend/src/main/java/team7/inplace/liked/likedInfluencer/domain/FavoriteInfluencer.java
@@ -1,11 +1,11 @@
-package team7.inplace.likedPlace.domain;
+package team7.inplace.liked.likedInfluencer.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import team7.inplace.place.domain.Place;
+import team7.inplace.influencer.domain.Influencer;
 import team7.inplace.user.domain.User;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
@@ -15,28 +15,23 @@ import static lombok.AccessLevel.PROTECTED;
 @RequiredArgsConstructor
 @NoArgsConstructor(access = PROTECTED)
 @Entity
-@Table(name = "liked_places")
-public class LikedPlace {
+public class FavoriteInfluencer {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
-
     @ManyToOne
     @NonNull
     @JoinColumn(name = "user_id")
     private User user;
-
     @ManyToOne
     @NonNull
-    @JoinColumn(name = "place_id")
-    private Place place;
-
+    @JoinColumn(name = "influencer_id")
+    private Influencer influencer;
     @Column
     private boolean isLiked = false;
 
     public void updateLike(boolean like) {
         this.isLiked = like;
     }
-
 }
