+++ backend/src/main/java/team7/inplace/liked/likedInfluencer/persistent/FavoriteInfluencerRepository.java
@@ -1,11 +1,11 @@
-package team7.inplace.likedInfluencer.persistent;
+package team7.inplace.liked.likedInfluencer.persistent;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import team7.inplace.likedInfluencer.domain.FavoriteInfluencer;
+import team7.inplace.liked.likedInfluencer.domain.FavoriteInfluencer;
 
 import java.util.List;
 import java.util.Optional;
