+++ backend/src/main/java/team7/inplace/place/application/PlaceService.java
@@ -17,8 +17,8 @@ import team7.inplace.global.exception.code.AuthorizationErrorCode;
 import team7.inplace.global.exception.code.PlaceErrorCode;
 import team7.inplace.global.exception.code.UserErrorCode;
 import team7.inplace.influencer.domain.Influencer;
-import team7.inplace.likedPlace.domain.LikedPlace;
-import team7.inplace.likedPlace.persistence.LikedPlaceRepository;
+import team7.inplace.liked.likedPlace.domain.LikedPlace;
+import team7.inplace.liked.likedPlace.persistence.LikedPlaceRepository;
 import team7.inplace.place.application.command.PlaceLikeCommand;
 import team7.inplace.place.application.command.PlacesCommand.Create;
 import team7.inplace.place.application.command.PlacesCommand.PlacesCoordinateCommand;
