+++ backend/src/main/java/team7/inplace/search/application/SearchService.java
@@ -10,10 +10,10 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import team7.inplace.likedInfluencer.persistent.FavoriteInfluencerRepository;
+import team7.inplace.liked.likedInfluencer.persistent.FavoriteInfluencerRepository;
 import team7.inplace.influencer.application.dto.InfluencerInfo;
 import team7.inplace.influencer.domain.Influencer;
-import team7.inplace.likedPlace.persistence.LikedPlaceRepository;
+import team7.inplace.liked.likedPlace.persistence.LikedPlaceRepository;
 import team7.inplace.search.application.dto.AutoCompletionInfo;
 import team7.inplace.search.application.dto.PlaceSearchInfo;
 import team7.inplace.search.application.dto.SearchType;
