+++ backend/src/test/java/team7/inplace/likedPlace/persistence/LikedPlaceRepositoryTest.java
@@ -13,7 +13,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import team7.inplace.global.exception.InplaceException;
 import team7.inplace.global.exception.code.LikedPlaceErrorCode;
-import team7.inplace.likedPlace.domain.LikedPlace;
+import team7.inplace.liked.likedPlace.domain.LikedPlace;
+import team7.inplace.liked.likedPlace.persistence.LikedPlaceRepository;
 import team7.inplace.place.domain.Place;
 import team7.inplace.user.domain.Role;
 import team7.inplace.user.domain.User;
